Task_1_Edge/requirements.txt
requests
paho-mqtt>=2.0.0